requirements.txt
aiologger
PySocks
requests
urwid
fabric